Revert "New manage/update"
