+++ src/__init__.py
@@ -10,4 +10,4 @@
 
 """
 
-__version__ = '0.0.6'
+__version__ = '0.0.5'
